Modify ilchz.cmnd file for gg generation
change the ilchz.cmnd file and the simulateall.sh script such that all Z decay channels are always
open and only the higgs decay channel in question is open. Previously only the Z decay channel with
the same decay products as the higgs decay was open. This lead to an error in the h->gg case.