Reconstruct K_s
in ilchz.cc:
Introduce two additional parameters defining a sperical shell around the interaction point. If a K_s decays into two charged pions inside this shellthe K_short is recorded. The two pions of this decay are not recorded. If the decay happens too close to the interaction point, the K_s is ignored and the pions may be recorded due to mistagging. If the decay happens too far away, the pions and the K_s are discarded.

in processhzroot.py:
add an additional parameter that allows to choose whether the selection of the two leading kaons is as before (i.e. ignores the K_s), considers additionaly the K_s, or considers only the K_s. The opposite_charge requirement is relaxed to demand the absolute value of the sum of the charge to be smaller than 2, i.e. allows +-, +0, -0 or 00 but not ++ or --