apply d0 cuts only to charged particles
hzplots.py now also deals (still in an unsophisticated way) with the K_s. Currently it basically
just excludes the K_s from the d0 cut (since it does not apply). Add the number of events per final
state (charged-charged, charged-neutral and neutral-neutral) to the plots of number of events