Fix bug of previous commit in hzplots leading to wrong results
By distinguishing different final states (KK, KsK, KsKs, etc.) the way the number of events was
calculated was wrong (efficiency*rate of events of this final state). With the newly introduced
choice of final states (and the corresponding cuts in the cut functions) and also different cuts for
different final states this is now replaced by just the efficiency, where the efficiency takes a cut
for the final state into accunt.

plottingall.sh is adjusted to loop over the final states