make scanning of efficiencies more automatic
make the generation of the legends and styles in hzplots automatically derived from the pid
point. Then hzplots does not need to be adjusted when the working points for the efficiencies are
changed.

plottingall extracts the available efficiency working points from the available root files and does
not re-generate the plots when the corresponding directory already exists.