Automatize the derivation of the pion reco efficiencies
The new kaonpionseparation script determines the efficiency of the pion mis-tag given a kaon reco
efficiency and a separation of the two gaussian distributions of some observable that distinguishes
between kaons and pions.
The script simulateall will run the kaonpionseparation script, automatically select the pion
efficiency and correctly name the output root file. From now on it takes some parameters for the
event generation