add smearing of impact parameter d0
smear the impact parameter d0 with a gaussian of width sqrt((0.005mm)^2 + sigma(d0)^2), where the
first contribution is the assumed resolution of the interaction point (reconstructed from Z decay)
and the second is the inherent resolution as a function of momentum and angle.

ATTENTION: The smearing happens in processhzroot.py, i.e. hzplots does not have access to true d0!

In hzplots, the shift of d0 by 1micron in the log-log-plot is removed since it is no longer
necessary.