Correct formula for impact parameter resolution
comparing the ILC equation in 1306.6329 with the CLIC equation and explanation in 1307.5288 I
conclude that the \oplus sign denotes adding in quadrature. I.e. the used formula is now

sigma(d0)=sqrt(5^2 + (10/(p*sin^(3/2)))^2)

instead of

sigma(d0)=     5   + 10/(p*sin^(3/2))

A dummy formula for the CLIC collider, replacing 10 by 15 is defined but commented out.